RAG-Project/requirements.txt
flask==2.3.3
flask-wtf==1.2.1
flask-bootstrap==3.3.7.1
python-dotenv==1.0.0
pinecone-client==3.0.0
langchain==0.0.327
langchain-google-genai==0.0.9
langchain-community==0.0.26
langchain-google-community==2.0.7
google-generativeai==0.3.1
bs4==0.0.1
requests==2.31.0
pytesseract==0.3.10
Pillow==10.0.0
pptx==0.6.22
python-docx==1.0.1
fpdf2==2.7.5
langchain-core==0.1.22
langchain-text-splitters==0.0.1
opencv-python==4.8.1.78
mediapipe==0.10.9
pygame==2.5.2
flask-session==0.5.0 